feat: Extract labelset forms into components to re-use in new task modal
