+++ lawnotation-ui/components/Labelsets.vue
@@ -1,0 +1,52 @@
+<template>
+  <div>
+    <div class="flex justify-between pt-2">
+      <h3 class="mb-2 text-lg font-semibold">Labelsets</h3>
+      <Button
+        label="Add"
+        icon="pi pi-plus"
+        icon-pos="right"
+        @click="$emit('addLabelset')"
+        data-test="create-new-labelset"
+      />
+    </div>
+    <Table
+      ref="labelsetTable"
+      endpoint="labelsets"
+      :sort="true"
+      :search="true"
+      :selectable="true"
+      @remove-rows="removeLabelsets"
+    >
+      <template #row="{ item }: { item: Labelset }">
+        <th scope="row" class="px-6 py-2 font-medium text-gray-900 whitespace-nowrap">
+          {{ item.id }}
+        </th>
+        <td class="px-6 py-2">
+          {{ item.name }}
+        </td>
+        <td class="px-6 py-2">
+          {{ item.desc }}
+        </td>
+        <td class="px-6 py-2 flex">
+          <Button label="View" size="small" @click="$emit('editLabelset', item.id)" data-test="edit-labelset" />
+        </td>
+      </template>
+    </Table>
+  </div>
+</template>
+
+<script setup lang="ts">
+import Table from "~/components/Table.vue";
+import type { Labelset } from "~/types";
+
+const { $trpc } = useNuxtApp();
+const labelsetTable = ref<InstanceType<typeof Table>>();
+
+const removeLabelsets = async (
+  ids: string[],
+  finish: (promises: Promise<Boolean>[]) => void
+) => {
+  finish([...ids.map((id) => $trpc.labelset.delete.mutate(+id))]);
+};
+</script>
